feat: changes in ticket component
